Time: 3 ms (56.48%), Space: 8.4 MB (61.57%) - LeetHub

+++ 0374-guess-number-higher-or-lower/0374-guess-number-higher-or-lower.c
@@ -16,7 +16,7 @@ int guessNumber(int n){
     while(low<=high)
     {
         int mid = low + (high-low)/2;
-        int result = guess((int)mid);
+        int result = guess(mid);
         if(result==0)
             return mid;
         else if(result==-1)
